feat(token-usage): fixed bucket counts, in-image titles/axes, smaller model labels
- Bucket granularity per range now: last hour 30s (120 bars), last day 15m (96),
  last week 2h (84), last month 6h (120). This required sub-minute buckets, so
  BucketMinutes is now a double end to end (query, summary result, builder's
  floor, REST parse, SDK, MCP tool) -- fractional widths like 0.5 are supported.
- The chart title and, for "Usage over time", the X ("Time") and Y ("Tokens")
  axis labels now render inside each SVG, so the Copy-image button includes them.
  The card header keeps just the copy button.
- Shrank the "Usage by model" row labels (9 -> 7.5) to match the time chart.

+++ src/Armada.Core/Models/TokenUsageQuery.cs
@@ -65,9 +65,10 @@ namespace Armada.Core.Models
         public int PageSize { get; set; } = 25;
 
         /// <summary>
-        /// Summary bucket width in minutes.
+        /// Summary bucket width in minutes. Fractional values are allowed (for example 0.5 for
+        /// 30-second buckets).
         /// </summary>
-        public int BucketMinutes { get; set; } = 15;
+        public double BucketMinutes { get; set; } = 15;
 
         /// <summary>
         /// Calculated zero-based offset.
